# ledger-stream/docs/openapi.yaml
openapi: 3.0.3
info:
  title: LedgerStream API
  description: >
    REST contract for LedgerStream. Settlement ingestion endpoint (/api/v1/settlements POST)
    and operational settlement projection query endpoints (/api/v1/settlements/{id} and
    /api/v1/settlements?accountId=...) are implemented and verified.
    LedgerStream records settlement events without moving money or calculating balances.
    Outbox dispatch, consumer idempotency and projection storage are verified.

  version: 1.0.0
  contact:
    name: Wellington Vanelli
    email: wvanelli@gmail.com
    url: https://github.com/w-vanelli/portfolio_pessoal/tree/main/ledger-stream

servers:
  - url: http://localhost:8080
    description: Local development environment

tags:
  - name: Settlements
    description: Settlement event ingestion and retrieval
  - name: Audit
    description: Immutable audit trail inspection for compliance and tracing
  - name: Operations
    description: Administrative and operational maintenance endpoints

paths:
  /api/v1/settlements:
    post:
      tags:
        - Settlements
      summary: Ingest settlement event with strict idempotency
      description: >
        Ingests a financial settlement record with an optional attachment. Enforces atomic idempotency
        via the `X-Idempotency-Key` header. Equivalent content includes metadata and attachment
        bytes. Cleanup is limited to confirmed pre-acceptance rollback; unknown commit outcomes
        require reconciliation. There is no cross-resource atomic transaction.
      parameters:
        - name: X-Idempotency-Key
          in: header
          required: true
          description: Unique client-generated key (UUID v4) guaranteeing idempotent execution.
          schema:
            type: string
            format: uuid
            example: "9b1deb4d-3b7d-4bad-9bdd-2b0d7b3dcb6d"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - metadata
              properties:
                metadata:
                  $ref: '#/components/schemas/SettlementIngestRequest'
                attachment:
                  type: string
                  format: binary
                  description: Optional binary receipt/invoice (PDF, PNG, JPG).
      responses:
        '201':
          description: Event durably accepted as COMMITTED; attachment promotion may remain pending.
          headers:
            Location:
              schema:
                type: string
              description: URI of the newly created settlement resource.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettlementResponse'
        '200':
          description: Idempotent replay - settlement was previously processed with this idempotency key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettlementResponse'
        '400':
          description: Invalid request payload or missing idempotency key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Idempotency conflict - payload checksum mismatch for the same key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Processing failure; acceptance outcome may require reconciliation. Retry with the same key and payload.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags:
        - Settlements
      summary: Query settlement projections by account identifier
      description: >
        Returns the operational projection of settlements for a specific account.
        The projection represents processed and accepted settlement events.
      parameters:
        - name: accountId
          in: query
          required: true
          description: Identifier of the merchant or tenant account.
          schema:
            type: string
            example: "ACC-US-9842"
      responses:
        '200':
          description: List of projected settlements for the account.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SettlementProjectionResponse'
        '400':
          description: Missing or invalid query parameter.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/settlements/{id}:
    get:
      tags:
        - Settlements
      summary: Retrieve settlement projection by identifier
      description: >
        Returns the operational projection of a settlement by its internal numeric settlement ID.
      parameters:
        - name: id
          in: path
          required: true
          description: Numeric internal settlement ID.
          schema:
            type: integer
            format: int64
            example: 1001
      responses:
        '200':
          description: Settlement projection details retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettlementProjectionResponse'
        '404':
          description: Settlement projection not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Invalid identifier.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'


  /api/v1/settlements/idempotency/{key}:
    get:
      tags:
        - Settlements
      summary: Query settlement by idempotency key
      parameters:
        - name: key
          in: path
          required: true
          description: The unique idempotency UUID.
          schema:
            type: string
            format: uuid
            example: "9b1deb4d-3b7d-4bad-9bdd-2b0d7b3dcb6d"
      responses:
        '200':
          description: Settlement found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SettlementResponse'
        '404':
          description: No settlement recorded under this key.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/settlements/{id}/audit:
    get:
      tags:
        - Audit
      summary: Retrieve settlement audit trail
      description: Returns the immutable lifecycle history of status changes and dispatch attempts.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Audit log history.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AuditLogEntry'

  /api/v1/admin/reconcile:
    post:
      tags:
        - Operations
      summary: Trigger on-demand reconciliation of orphaned staging files, pending promotions, and outbox integrity
      description: >
        Executes a full reconciliation cycle immediately: purges unreferenced orphaned staging files
        exceeding the grace period, regularizes STAGED attachments with committed events, and
        reconstructs missing outbox entries or aligns event/outbox statuses. Returns a detailed report.
      responses:
        '200':
          description: Reconciliation completed. Returns detailed counts and any isolated item failures.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReconciliationReport'

components:
  schemas:
    SettlementIngestRequest:
      type: object
      required:
        - accountId
        - currency
        - amount
        - settlementType
      properties:
        accountId:
          type: string
          example: "ACC-US-9842"
          description: Identifier of the merchant or tenant account.
        currency:
          type: string
          example: "USD"
          minLength: 3
          maxLength: 3
        amount:
          type: string
          example: "1450.75"
          pattern: '^(0|[1-9]\d{0,12})(\.\d{1,2})?$'
          description: >
            Monetary amount as a string decimal. Must be strictly positive (> 0).
            Permitted range: 0.01 to 9,999,999,999,999.99 (NUMERIC(15,2)).
            Direct String-to-BigDecimal conversion prevents floating-point precision loss.
            Representations "1", "1.0", and "1.00" are equivalent after normalization.
            The regex pattern is a format guard; the domain enforces positivity and range.
        settlementType:
          type: string
          enum: [CARD_PAYOUT, WIRE_TRANSFER, INVOICE_SETTLEMENT, CHARGEBACK_ADJUSTMENT]
          example: "INVOICE_SETTLEMENT"
        originalSettlementId:
          type: integer
          format: int64
          nullable: true
          description: Required for CHARGEBACK_ADJUSTMENT; references an existing non-chargeback event of the same account and currency. Forbidden for other types.
        description:
          type: string
          example: "Monthly recurring SaaS license payout batch #84"
          maxLength: 255

    SettlementProjectionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1001
        settlementId:
          type: integer
          format: int64
          example: 1001
        messageId:
          type: string
          format: uuid
          example: "9b1deb4d-3b7d-4bad-9bdd-2b0d7b3dcb6d"
        accountId:
          type: string
          example: "ACC-US-9842"
        currency:
          type: string
          example: "USD"
        amount:
          type: string
          example: "1450.75"
        settlementType:
          type: string
          example: "WIRE_TRANSFER"
        originalSettlementId:
          type: integer
          format: int64
          nullable: true
          example: null
        acceptedAt:
          type: string
          format: date-time
        processedAt:
          type: string
          format: date-time

    SettlementResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1001

        idempotencyKey:
          type: string
          format: uuid
          example: "9b1deb4d-3b7d-4bad-9bdd-2b0d7b3dcb6d"
        accountId:
          type: string
          example: "ACC-US-9842"
        currency:
          type: string
          example: "USD"
        amount:
          type: string
          example: "1450.75"
        status:
          type: string
          description: COMMITTED is durable acceptance, even if files/dispatch remain pending. Future DISPATCHED requires confirmed publication, required routing and ready attachments; it does not imply consumption or settlement.
          enum: [STAGED, COMMITTED, DISPATCHED, FAILED, COMPENSATED]
          example: "COMMITTED"
        hasAttachment:
          type: boolean
          example: true
        attachmentFileName:
          type: string
          example: "invoice-84.pdf"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    AuditLogEntry:
      type: object
      properties:
        id:
          type: integer
          format: int64
        settlementId:
          type: integer
          format: int64
        previousStatus:
          type: string
          example: "STAGED"
        newStatus:
          type: string
          example: "COMMITTED"
        eventDetails:
          type: string
          example: "Transactional commit completed; file promoted from staging to permanent storage."
        timestamp:
          type: string
          format: date-time

    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 400
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "Missing required header: X-Idempotency-Key"
        path:
          type: string
          example: "/api/v1/settlements"

    ReconciliationReport:
      type: object
      description: Summary of an executed reconciliation cycle.
      properties:
        startedAt:
          type: string
          format: date-time
          example: "2026-10-01T20:18:39.023Z"
        completedAt:
          type: string
          format: date-time
          example: "2026-10-01T20:18:39.068Z"
        orphanDirectoriesRemoved:
          type: integer
          example: 0
          description: Number of unreferenced staging attempt directories older than grace period that were deleted.
        attachmentsReconciled:
          type: integer
          example: 0
          description: Number of STAGED attachments whose permanent files were regularized or purged.
        outboxRepairs:
          type: integer
          example: 0
          description: Number of outbox rows reconstructed or status-aligned.
        failures:
          type: array
          items:
            type: string
          example: []
          description: List of isolated item error messages encountered during the cycle.